Backend_v2/requirements.txt
# Web 框架
quart~=0.20.0
quart-cors~=0.8.0
quart-jwt-extended~=0.1.0
werkzeug<4.0.0

# 异步支持
aiohttp~=3.11.16
asyncpg~=0.30.0

# 任务调度
APScheduler~=3.11.0

# 环境与配置
python-dotenv~=1.1.0

# 工具库
PyJWT~=2.10.1
requests~=2.32.3

# 生产环境
hypercorn~=0.17.3

# 加密支持
pycryptodome~=3.22.0

# 类型提示
pydantic~=2.11.3
pydantic-settings~=2.8.1

# 数据保存
PyYAML~=6.0.2
cryptography~=39.0.1